# Fix MemoryIndex concurrency issue