When iterating over commits and computing patches concurrently within goroutines, the go race detector detects concurrent writes and reads to the offsetHash map:

```
fatal error: concurrent map read and map write
goroutine 13 [running]:
runtime.throw(0x8af5e9, 0x21)
	/usr/local/go/src/runtime/panic.go:1116 +0x72 fp=0xc00050b540 sp=0xc00050b510 pc=0x437672
runtime.mapaccess2_fast64(0x81eae0, 0xc00017f650, 0x23a43, 0x40d8b0, 0xc0003144c0)
	/usr/local/go/src/runtime/map_fast64.go:61 +0x1ac fp=0xc00050b568 sp=0xc00050b540 pc=0x412a2c
github.com/go-git/go-git/v5/plumbing/format/idxfile.(*MemoryIndex).FindHash(0xc000192000, 0x23a43, 0x0, 0x0, 0x7f3500000000, 0x0, 0xc00031c120)
	/builds/.cache/pkg/mod/github.com/go-git/go-git/v5@v5.1.0/plumbing/format/idxfile/idxfile.go:176 +0x1a5 fp=0xc00050b5c8 sp=0xc00050b568 pc=0x62f8e5
```

Fixed by adding a sync.RWMutex to be used during reads and writes.

Thank you for the excellent lib.