@dustin-decker 2020-11-25
I prefer https://github.com/go-git/go-git/pull/186 over this fix but would be very happy if either get merged :pray: