@cep21 2021-11-10
Thanks for this.  It was very surprising to me that even read only operations on go-git are not race safe.