@pjbgf 2022-12-03
I ran the existing benchmarks on this (rebased), and here are the results:

```
goos: linux goarch: amd64
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz

name           old time/op    new time/op    delta
FindOffset-16     279ns ± 4%     427ns ± 7%   ~     (p=0.100 n=3+3)
FindCRC32-16      181ns ± 2%     217ns ±15%   ~     (p=0.100 n=3+3)
Contains-16       124ns ± 1%     142ns ±12%   ~     (p=0.100 n=3+3)
Entries-16       1.23µs ± 4%    0.97µs ±19%   ~     (p=0.100 n=3+3)
```

I would prefer this fix over #186.