@pjbgf 2023-05-20
I created a branch to start amassing fixes around concurrency issues. As this has a performance cost, I won't propose merging it into `master` just yet. Folks insterested on using it can consume those changes via:

`go get github.com/go-git/go-git/v5@read-concurrency`

Closing this in favour of the new branch.